refactor & fix numeraires update bug
