+++ apps/extension/src/routes/page/onboarding/set-numeraire.tsx
@@ -2,7 +2,7 @@ import { Card, CardDescription, CardHeader, CardTitle } from '@penumbra-zone/ui/
 import { FadeTransition } from '@penumbra-zone/ui/components/ui/fade-transition';
 import { usePageNav } from '../../../utils/navigate';
 import { PagePath } from '../paths';
-import { NumeraireForm } from '../../../shared/components/numeraires';
+import { NumeraireForm } from '../../../shared/components/numeraires-form';
 
 export const SetNumerairesPage = () => {
   const navigate = usePageNav();
@@ -16,14 +16,14 @@ export const SetNumerairesPage = () => {
       <Card className='w-[400px]' gradient>
         <CardHeader>
           <CardTitle>In which token denomination would you prefer to price assets?</CardTitle>
+          <CardDescription>
+            Prax does not use third-party price providers for privacy reasons, instead Prax indexes
+            asset prices locally by selected denomination
+          </CardDescription>
         </CardHeader>
-
-        <CardDescription>
-          Prax does not use third-party price providers for privacy reasons, instead Prax indexes
-          asset prices locally by selected denomination
-        </CardDescription>
-
+        <div className='mt-6'>
           <NumeraireForm isOnboarding onSuccess={onSuccess} />
+        </div>
       </Card>
     </FadeTransition>
   );
