+++ apps/extension/src/shared/components/numeraires-form.tsx
@@ -1,13 +1,13 @@
-import { SelectList } from '@penumbra-zone/ui/components/ui/select-list';
 import { ChainRegistryClient } from '@penumbra-labs/registry';
-import { AllSlices } from '../../../state';
-import { useStoreShallow } from '../../../utils/use-store-shallow';
-import { FormEvent, useMemo } from 'react';
-import { Button } from '@penumbra-zone/ui/components/ui/button';
-import { Metadata } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb';
-import { useChainIdQuery } from '../../../hooks/chain-id';
+import { AllSlices, useStore } from '../../state';
+import { useChainIdQuery } from '../../hooks/chain-id';
+import { useMemo, useState } from 'react';
+import { ServicesMessage } from '../../message/services';
+import { SelectList } from '@penumbra-zone/ui/components/ui/select-list';
 import { bech32mAssetId } from '@penumbra-zone/bech32m/passet';
 import { getAssetId } from '@penumbra-zone/getters/metadata';
+import { Button } from '@penumbra-zone/ui/components/ui/button';
+import { Metadata } from '@buf/penumbra-zone_penumbra.bufbuild_es/penumbra/core/asset/v1/asset_pb';
 
 const getNumeraireFromRegistry = (chainId: string): Metadata[] => {
   const registryClient = new ChainRegistryClient();
@@ -23,34 +23,39 @@ const useNumerairesSelector = (state: AllSlices) => {
   };
 };
 
-
-
-export const NumeraireForm = ({ isOnboarding, onSuccess }: { isOnboarding?: boolean, onSuccess: () => void | Promise<void>;}) => {
+export const NumeraireForm = ({
+  isOnboarding,
+  onSuccess,
+}: {
+  isOnboarding?: boolean;
+  onSuccess: () => void | Promise<void>;
+}) => {
   const { chainId } = useChainIdQuery();
-  const { selectedNumeraires, selectNumeraire, saveNumeraires } =
-    useStoreShallow(useNumerairesSelector);
+  const { selectedNumeraires, selectNumeraire, saveNumeraires } = useStore(useNumerairesSelector);
   const numeraires = useMemo(
-    () => getNumeraireFromRegistry(chainId || 'penumbra-testnet-deimos-8'),
+    () => getNumeraireFromRegistry(chainId ?? 'penumbra-testnet-deimos-8'),
     [chainId],
   );
 
-  const onSubmit = async (
-      onSuccess: () => void | Promise<void>,
-  ) => {
-    saveNumeraires();
-    await onSuccess();
+  const [loading, setLoading] = useState(false);
+
+  const handleSubmit = () => {
+    setLoading(true);
+    void (async function () {
+      await saveNumeraires();
+      await chrome.runtime.sendMessage(ServicesMessage.ChangeNumeraires);
+      await onSuccess();
+    })();
   };
 
-  const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
-    e.preventDefault();
-    void onSubmit(onSuccess);
-  };
   return (
     <>
       <div className='flex flex-col gap-2'>
         <form className='flex flex-col gap-4' onSubmit={handleSubmit}>
           <SelectList>
             {numeraires.map(metadata => {
+              // Image default is "" and thus cannot do nullish-coalescing
+              // eslint-disable-next-line @typescript-eslint/prefer-nullish-coalescing
               const icon = metadata.images[0]?.png || metadata.images[0]?.svg;
               return (
                 <SelectList.Option
@@ -76,9 +81,11 @@ export const NumeraireForm = ({ isOnboarding, onSuccess }: { isOnboarding?: bool
               className='my-5'
               key='save-button'
               variant='gradient'
-              type= 'submit'
+              type='submit'
+              disabled={loading}
+              onClick={handleSubmit}
             >
-              {isOnboarding ? 'Next' : 'Save'}
+              {isOnboarding ? 'Next' : loading ? 'Saving...' : 'Save'}
             </Button>
           </SelectList>
         </form>
